add GCPS in rasterio.control and update tests
